Move redux connect to manage tokens scene
